pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "object-detection"
version = "2.0.0"
description = "Production-quality object detection system for tracking movement"
authors = [{name = "Shane"}]
license = {text = "MIT"}
requires-python = ">=3.7"
dependencies = [
    "ultralytics>=8.0.0",
    "opencv-python>=4.5.0",
    "pyyaml>=6.0",
    "numpy>=1.21.0",
    "pydantic>=2.0.0",
    "questionary>=2.0.0",
    "requests>=2.28.0",
]

[project.scripts]
object-detection = "object_detection.cli:main"

[tool.setuptools.package-data]
object_detection = ["*.yaml"]

[tool.setuptools.packages.find]
where = ["src"]
